# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Socious-API
  version: 2.2.6
servers:
  - url: https://dev.socious.io/api/v3/
    description: Development server (uses test data)
  - url: https://socious.io/api/v3/
    description: Production server (uses live data)
  - url: http://localhost:3000
    description: Local server (for development purposes)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    JobCategories:
      properties:
        createdAt:
          format: date-time
          type: string
        hourlyWageDollars:
          type: number
        id:
          type: string
        name:
          type: string
        projects:
          items:
            $ref: '#/components/schemas/Projects'
          type: array
        updatedAt:
          format: date-time
          type: string
      required:
        - createdAt
        - hourlyWageDollars
        - id
        - name
        - projects
        - updatedAt
      type: object
    Projects:
      properties:
        # answers:
        #   items:
        #     $ref: '#/components/schemas/Answers'
        #   type: array
        # applicants:
        #   $ref: '#/components/schemas/Applicants'
        city:
          type: string
        commitmentHoursHigher:
          type: string
        commitmentHoursLower:
          type: string
        country:
          type: string
        createdAt:
          format: date-time
          type: string
        deletedAt:
          format: date-time
          type: string
        description:
          type: string
        # escrows:
        #   items:
        #     $ref: '#/components/schemas/Escrows'
        #   type: array
        experienceLevel:
          type: number
        expiresAt:
          format: date-time
          type: string
        # feedbacks:
        #   items:
        #     $ref: '#/components/schemas/Feedbacks'
        #   type: array
        # geoname:
        #   $ref: '#/components/schemas/Geonames'
        id:
          type: string
        # identity:
        #   $ref: '#/components/schemas/Identities'
        jobCategory:
          $ref: '#/components/schemas/JobCategories'
        # missions:
        #   items:
        #     $ref: '#/components/schemas/Missions'
        #   type: array
        # offers:
        #   items:
        #     $ref: '#/components/schemas/Offers'
        #   type: array
        oldId:
          type: number
        otherPartyId:
          type: string
        otherPartyTitle:
          type: string
        otherPartyUrl:
          type: string
        paymentCurrency:
          type: string
        paymentRangeHigher:
          type: string
        paymentRangeLower:
          type: string
        paymentScheme:
          enum:
            - FIXED
            - HOURLY
          type: string
        paymentType:
          enum:
            - PAID
            - VOLUNTEER
          type: string
        projectLength:
          enum:
            - 1_3_MONTHS
            - 3_6_MONTHS
            - 6_MONTHS_OR_MORE
            - LESS_THAN_A_DAY
            - LESS_THAN_A_MONTH
          type: string
        projectType:
          enum:
            - FULL_TIME
            - ONE_OFF
            - PART_TIME
          type: string
        # questions:
        #   items:
        #     $ref: '#/components/schemas/Questions'
        #   type: array
        remotePreference:
          enum:
            - HYBRID
            - ONSITE
            - REMOTE
          type: string
        searchTsv:
          type: string
        skills:
          items:
            type: string
          type: array
        status:
          enum:
            - ACTIVE
            - DRAFT
            - EXPIRE
          type: string
        # submittedWorks:
        #   items:
        #     $ref: '#/components/schemas/SubmittedWorks'
        #   type: array
        title:
          type: string
        updatedAt:
          format: date-time
          type: string
        weeklyHoursHigher:
          type: string
        weeklyHoursLower:
          type: string
        project_kind:
          enum:
            - JOB
            - SERVICE
          type: string
        service_length:
          enum:
            - LESS_THAN_A_DAY
            - 1_3_DAYS
            - 1_WEEK
            - 2_WEEKS
            - 1_MONTH
            - 1_3_MONTHS
            - 3_6_MONTHS
            - 6_MONTHS_OR_MORE
          type: string
        service_total_hours:
          type: number
        service_price:
          type: number
        service_work_samples:
          items:
            properties:
              file:
                type: string
              url:
                type: string
            type: object
          type: array
      required:
        - answers
        - applicants
        - city
        - commitmentHoursHigher
        - commitmentHoursLower
        - country
        - createdAt
        - deletedAt
        - description
        - escrows
        - experienceLevel
        - expiresAt
        - feedbacks
        - geoname
        - id
        - identity
        - jobCategory
        - missions
        - offers
        - oldId
        - otherPartyId
        - otherPartyTitle
        - otherPartyUrl
        - paymentCurrency
        - paymentRangeHigher
        - paymentRangeLower
        - paymentScheme
        - paymentType
        - projectLength
        - projectType
        - questions
        - remotePreference
        - searchTsv
        - skills
        - status
        - submittedWorks
        - title
        - updatedAt
        - weeklyHoursHigher
        - weeklyHoursLower
      type: object
    Contracts:
      properties:
        id:
          type: string
          example: 282bd9ef-73cf-4c4c-bcf0-09615930d408
        name:
          type: string
          example: name
        description:
          type: string
          example: Description of the Contract
        type:
          enum:
            - VOLUNTEER
            - PAID
          type: string
        total_amount:
          type: number
          example: 12.5
        currency:
          enum:
            - USD
            - JPY
          type: string
        crypto_currency:
          type: string
        currency_rate:
          type: number
          example: 12.5
        commitment:
          type: number
          example: 1
        commitment_period:
          enum:
            - HOURLY
            - DAILY
            - WEEKLY
            - MONTHLY
          type: string
        commitment_period_count:
          type: number
          example: 1
        payment_type:
          enum:
            - CRYPTO
            - FIAT
          type: string
        applicant_id:
          type: string
          example: 3a7a85d4-e4f0-4b98-9922-5e5aa9456a6b
        project_id:
          type: string
          example: 3a7a85d4-e4f0-4b98-9922-5e5aa9456a6b
        client_id:
          type: string
          example: 3a7a85d4-e4f0-4b98-9922-5e5aa9456a6b
        status:
          enum:
            - CREATED
            - CLIENT_APPROVED
            - SIGNED
            - PROVIDER_CANCELED
            - CLIENT_CANCELED
            - APPLIED
            - COMPLETED
          type: string
        provider_id:
          type: string
          example: 3a7a85d4-e4f0-4b98-9922-5e5aa9456a6b
        payment_id:
          type: string
          example: 3a7a85d4-e4f0-4b98-9922-5e5aa9456a6b
        offer_id:
          type: string
          example: 3a7a85d4-e4f0-4b98-9922-5e5aa9456a6b
        mission_id:
          type: string
          example: 3a7a85d4-e4f0-4b98-9922-5e5aa9456a6b
        created_at:
          format: date-time
          type: string
        updated_at:
          format: date-time
          type: string
paths:
  /services:
    get:
      summary: Get All Services
      security:
        - bearerAuth: []
      description: Get All Services
      tags:
        - Services
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Projects'
                  total:
                    type: number
                type: object
    post:
      summary: Create a Service
      security:
        - bearerAuth: []
      description: Create a Service
      tags:
        - Services
      requestBody:
        content:
          application/json:
            schema:
              properties:
                title:
                  type: string
                  example: Title of the Service
                description:
                  type: string
                  example: Description of the Service
                payment_currency:
                  type: string
                skills:
                  type: array
                  items:
                    type: string
                  example:
                    - Skill1
                    - Skill2
                    - Skill3
                job_category_id:
                  type: string
                  example: 282bd9ef-73cf-4c4c-bcf0-09615930d408
                service_total_hours:
                  type: number
                  example: 100
                service_price:
                  type: string
                  example: 2000
                service_length:
                  enum:
                    - LESS_THAN_A_DAY
                    - 1_3_DAYS
                    - 1_WEEK
                    - 2_WEEKS
                    - 1_MONTH
                    - 1_3_MONTHS
                    - 3_6_MONTHS
                    - 6_MONTHS_OR_MORE
                  type: string
                work_samples:
                  type: array
                  items:
                    type: string
                  example:
                    - 1744f7d6-d90a-4e35-acc1-e59d5a18e925
                    - 03fa3d0d-9204-4cea-b6d7-7858b72bfee5
                    - 3a7a85d4-e4f0-4b98-9922-5e5aa9456a6b
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Projects'
  /services/:id:
    parameters:
      - name: id
        required: true
        in: path
        description: Service ID
        schema:
          type: string
    get:
      summary: Get Service
      security:
        - bearerAuth: []
      description: Get Service
      tags:
        - Services
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Projects'
    patch:
      summary: Update Service
      security:
        - bearerAuth: []
      description: Update Service
      tags:
        - Services
      requestBody:
        content:
          application/json:
            schema:
              properties:
                title:
                  type: string
                  example: Title of the Service
                description:
                  type: string
                  example: Description of the Service
                payment_currency:
                  type: string
                skills:
                  type: array
                  items:
                    type: string
                  example:
                    - Skill1
                    - Skill2
                    - Skill3
                job_category_id:
                  type: string
                  example: 282bd9ef-73cf-4c4c-bcf0-09615930d408
                service_total_hours:
                  type: number
                  example: 100
                service_price:
                  type: string
                  example: 2000
                service_length:
                  enum:
                    - LESS_THAN_A_DAY
                    - 1_3_DAYS
                    - 1_WEEK
                    - 2_WEEKS
                    - 1_MONTH
                    - 1_3_MONTHS
                    - 3_6_MONTHS
                    - 6_MONTHS_OR_MORE
                  type: string
                work_samples:
                  type: array
                  items:
                    type: string
                  example:
                    - 1744f7d6-d90a-4e35-acc1-e59d5a18e925
                    - 03fa3d0d-9204-4cea-b6d7-7858b72bfee5
                    - 3a7a85d4-e4f0-4b98-9922-5e5aa9456a6b
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Projects'
    delete:
      summary: Delete Service
      security:
        - bearerAuth: []
      description: Delete Service
      tags:
        - Services
      responses:
        '200':
          description: Successful response
          content:
            application/json: 
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: success
  /contracts:
    get:
      summary: Get all contracts
      security:
        - bearerAuth: []
      description: Get all contracts
      tags:
        - Contracts
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contracts'
    post:
      summary: Create contract
      security:
        - bearerAuth: []
      description: Create contract
      tags:
        - Contracts
      requestBody:
        content:
          application/json:
            schema:
              properties:
                title:
                  type: string
                  example: Title of the Contract
                description:
                  type: string
                  example: Description of the Contract
                type:
                  enum:
                    - VOLUNTEER
                    - PAID
                  type: string
                total_amount:
                  type: number
                  example: 12.5
                currency:
                  enum:
                    - USD
                    - JPY
                  type: string
                crypto_currency:
                  type: string
                currency_rate:
                  type: number
                  example: 12.5
                commitment:
                  type: number
                  example: 1
                commitment_period:
                  enum:
                    - HOURLY
                    - DAILY
                    - WEEKLY
                    - MONTHLY
                  type: string
                commitment_period_count:
                  type: number
                  example: 1
                payment_type:
                  enum:
                    - CRYPTO
                    - FIAT
                  type: string
                applicant_id:
                  type: string
                  example: 3a7a85d4-e4f0-4b98-9922-5e5aa9456a6b
                project_id:
                  type: string
                  example: 3a7a85d4-e4f0-4b98-9922-5e5aa9456a6b
                client_id:
                  type: string
                  example: 3a7a85d4-e4f0-4b98-9922-5e5aa9456a6b
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contracts'
  /contracts/:id:
    parameters:
        - name: id
          required: true
          in: path
          description: Contract ID
          schema:
            type: string
    get:
      summary: Get contract
      security:
        - bearerAuth: []
      description: Get contract
      tags:
        - Contracts
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contracts'
    patch:
      summary: Update contract
      security:
        - bearerAuth: []
      description: Update contract
      tags:
        - Contracts
      requestBody:
        content:
          application/json:
            schema:
              properties:
                title:
                  type: string
                  example: Title of the Contract
                description:
                  type: string
                  example: Description of the Contract
                type:
                  enum:
                    - VOLUNTEER
                    - PAID
                  type: string
                total_amount:
                  type: number
                  example: 12.5
                currency:
                  enum:
                    - USD
                    - JPY
                  type: string
                crypto_currency:
                  type: string
                currency_rate:
                  type: number
                  example: 12.5
                commitment:
                  type: number
                  example: 1
                commitment_period:
                  enum:
                    - HOURLY
                    - DAILY
                    - WEEKLY
                    - MONTHLY
                  type: string
                commitment_period_count:
                  type: number
                  example: 1
                payment_type:
                  enum:
                    - CRYPTO
                    - FIAT
                  type: string
                applicant_id:
                  type: string
                  example: 3a7a85d4-e4f0-4b98-9922-5e5aa9456a6b
                project_id:
                  type: string
                  example: 3a7a85d4-e4f0-4b98-9922-5e5aa9456a6b
                client_id:
                  type: string
                  example: 3a7a85d4-e4f0-4b98-9922-5e5aa9456a6b
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contracts'
  /contracts/:id/deposit:
    parameters:
        - name: id
          required: true
          in: path
          description: Contract ID
          schema:
            type: string
    post:
      summary: Deposit on contract
      security:
        - bearerAuth: []
      description: Deposit on contract
      tags:
        - Contracts
      requestBody:
        content:
          application/json:
            schema:
              properties:
                card_id:
                  type: string
                  example: The selected card to deposit (fiat)
                txid:
                  type: string
                  example: The transaction hash (crypto)
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contracts'